pyproject.toml
[tool.poetry]
name = "untitled"
version = "0.1.0"
description = "A Community Bot."
authors = ["Arnav"]
license = "MIT"
readme = "README.md"
homepage = ""
repository = "https://github.com/Arnav-2004/untitled-bot"
packages = [
    { include = "bot" },
]

[tool.poetry.dependencies]
python = "3.9.*"
"discord.py" = {url = "https://github.com/Rapptz/discord.py/archive/45d498c1b76deaf3b394d17ccf56112fa691d160.zip"}
humanize = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "~=21.9b0"
flake8 = "~=3.9.2"
flake8-annotations = "~=2.6.2"
flake8-bugbear = "~=21.9.1"
flake8-docstrings = "~=1.6.0"
flake8-import-order = "~=0.18.1"
flake8-string-format = "~=0.3.0"
flake8-tidy-imports = "~=4.4.1"
flake8-todo = "~=0.7"
pep8-naming = "~=0.12.1"
pre-commit = "~=2.15.0"
python-dotenv = "~=0.19.0"
taskipy = "~=1.8.2"

[tool.taskipy.tasks]
bot = { cmd = "python -m bot", help = "Runs the Bot"}
lint = { cmd = "pre-commit run --all-files", help = "Lints the project" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit git hook" }
format = { cmd = "black --check .", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
